clean app service funcs
